Organiza estrutura de arquivos e modelos,
