--- autismcare/authentication/models/__init__.py
@@ -1,6 +1,0 @@
-from .base import AutenticationUser, Address
-from .role import Role
-from .employee import Employee
-from .parent import Parent
-from .child import Child
-from .medical import Allergy, MedicalRecord, MedicalHistory
