--- autismcare/authentication/models/child.py
@@ -1,32 +1,0 @@
-from django.db import models
-from .parent import Parent
-
-class Child(models.Model):
-    GENDER_CHOICES = [
-        ('M', 'Masculino'),
-        ('F', 'Feminino'),
-        ('O', 'Outro'),
-    ]
-
-    parent = models.ForeignKey(
-        Parent,
-        on_delete=models.CASCADE,
-        related_name="children"
-    )
-    first_name = models.CharField(max_length=100)
-    last_name = models.CharField(max_length=100)
-    birthdate = models.DateField(verbose_name="Data de Nascimento")
-    gender = models.CharField(
-        max_length=1,
-        choices=GENDER_CHOICES,
-        blank=True,
-        null=True
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return f"{self.first_name} {self.last_name} (Responsável: {self.parent.user.get_full_name()})"
-
-    class Meta:
-        indexes = [models.Index(fields=['parent'])]
