--- autismcare/authentication/models/employee.py
@@ -1,40 +1,0 @@
-from django.db import models
-from django.core.validators import MinValueValidator
-from .base import AutenticationUser
-from .role import Role
-
-class Employee(models.Model):
-    STATUS_CHOICES = [
-        ('active', 'Ativo'),
-        ('inactive', 'Inativo'),
-        ('terminated', 'Demitido'),
-    ]
-
-    user = models.OneToOneField(
-        AutenticationUser,
-        on_delete=models.CASCADE,
-        related_name='employee'
-    )
-    role = models.ForeignKey(
-        Role,
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        verbose_name="Função"
-    )
-    salary = models.DecimalField(
-        max_digits=10,
-        decimal_places=2,
-        validators=[MinValueValidator(0)]
-    )
-    admission_date = models.DateField()
-    status = models.CharField(
-        max_length=20,
-        choices=STATUS_CHOICES,
-        default='active'
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return f"{self.user.username} - {self.role.name if self.role else 'Sem função'}"
