--- autismcare/authentication/models/role.py
@@ -1,28 +1,0 @@
-from django.db import models
-from django.contrib.auth.models import Permission
-
-class Role(models.Model):
-    name = models.CharField(max_length=50, unique=True, verbose_name="Nome da Role")
-    description = models.TextField(blank=True, null=True, verbose_name="Descrição")
-    permissions = models.ManyToManyField(
-        Permission,
-        blank=True,
-        verbose_name="Permissões"
-    )
-    parent_role = models.ForeignKey(
-        'self',
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        verbose_name="Role Superior"
-    )
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    def __str__(self):
-        return self.name
-
-    class Meta:
-        verbose_name = "Role"
-        verbose_name_plural = "Roles"
-        ordering = ["name"]
